% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{login}
\alias{login}
\title{Log in to the local neo4j server}
\usage{
login(con = NULL)
}
\arguments{
\item{con}{an existed connexion object. It is not necessary to log in
for the first time.}
}
\value{
connection to the local neo4j database
}
\description{
Before running \code{login()}, you have to successfully finish the Reactome Neo4j
database setup and build a connection on your local machine (details see:
https://github.com/reactome/ReactomeGraph4R). This command is to create a
neo4r object that is used to communicate between R and Neo4j, also to do a
sanity check for the connection.
}
\examples{
\dontrun{
# The first step to the graph database!
login()

# you can also check the neo4r connexion object by running:
getOption("con")
}
}
