% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchDiseases}
\alias{matchDiseases}
\title{MATCH diseases of PhysicalEntity/Reaction/Pathway}
\usage{
matchDiseases(
  id = NULL,
  displayName = NULL,
  species = NULL,
  type = c("row", "graph")
)
}
\arguments{
\item{id}{stId or dbId of a PhysicalEntity/Event/Disease}

\item{displayName}{displayName of a PhysicalEntity/Event/Disease}

\item{species}{name or taxon id or dbId or abbreviation of the specified species}

\item{type}{return results as a list of dataframes (\strong{'row'}) or as a graph object (\strong{'graph'})}
}
\value{
Disease(s) related to the given PhysicalEntity/Reaction/Pathway; or instances related to the given Disease
}
\description{
Find Diseases related to a PhysicalEntity or an Event or
get PhysicalEntities/Events associated with a Disease in reverse
}
\examples{
matchDiseases(displayName="neuropathy", species="M. musculus", type="row")
matchDiseases(id="R-HSA-162588", type="graph")
}
\seealso{
Other match: 
\code{\link{matchHierarchy}()},
\code{\link{matchObject}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
