% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchEventPaths}
\alias{matchEventPaths}
\title{MATCH the preceding/following Events}
\usage{
matchEventPaths(
  id = NULL,
  name = NULL,
  species = NULL,
  depth = 1,
  all.depth = FALSE,
  type = c("row", "graph")
)
}
\arguments{
\item{id}{a stable or db id of an Event}

\item{name}{an Event name}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{depth}{number of depths}

\item{all.depth}{if set to \code{TRUE}, all RLEs connected to the given Event in all depths returned}

\item{type}{to return results as a list of dataframes (\strong{'row'}) or as a graph object (\strong{'graph'})}
}
\value{
preceding/following Events connected to the given Event in specified depth(s), default depth = 1
}
\description{
MATCH the preceding/following Events
}
\examples{
matchEventPaths("R-HSA-983150", all.depth=TRUE, type="row")
}
\seealso{
Other match: 
\code{\link{matchHierarchy}()},
\code{\link{matchObject}()},
\code{\link{matchReactionsInPathway}()}
}
\concept{match}
