% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchInteractors}
\alias{matchInteractors}
\title{MATCH interactors}
\usage{
matchInteractors(
  pe.id = NULL,
  pe.displayName = NULL,
  species = NULL,
  type = c("row", "graph")
)
}
\arguments{
\item{pe.id}{stId or dbId of a PhysicalEntity}

\item{pe.displayName}{displayName of a PhysicalEntity}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{type}{return results as a list of dataframes (\strong{'row'}),
or as a graph object (\strong{'graph'})}
}
\value{
interactions of a given PhysicalEntity
}
\description{
To retrieve interactions of a given PhysicalEntity (PE), it first finds the
ReferenceEntity matched with the PE, then get the Interactions having
"interactor" relationship with the ReferenceEntity.
}
\examples{
pe.id <- 996766
# matchInteractors(pe.id)
}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchObject}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
