% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchObject}
\alias{matchObject}
\title{Basic query for database objects}
\usage{
matchObject(
  id = NULL,
  displayName = NULL,
  schemaClass = NULL,
  species = NULL,
  returnedAttributes = NULL,
  property = NULL,
  relationship = NULL,
  limit = NULL,
  databaseName = "Reactome"
)
}
\arguments{
\item{id}{Reactome stId or dbId, or non-Reactome identifier}

\item{displayName}{displayName of a database object}

\item{schemaClass}{schema class of a database object}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{returnedAttributes}{specific attribute(s) to be returned.
If set to \code{NULL}, all attributes returned}

\item{property}{a list of property keys and values,
e.g. \code{list(isChimeric = TRUE, isInDisease = TRUE)}}

\item{relationship}{a relationship type}

\item{limit}{the number of returned objects}

\item{databaseName}{database name. All databases see
\href{https://reactome.org/content/schema/objects/ReferenceDatabase}{here}}
}
\value{
Reactome database object(s) that meets all specified conditions
}
\description{
This function can fetch instance by setting arguments:
\itemize{
\item \strong{id}: a Reactome dbId/stId, or non-Reactome id
\item \strong{displayName}: a display name of a Reactome object
\item \strong{schemaClass}: a specific schema class,
see \href{https://reactome.org/content/schema/DatabaseObject}{Data Schema}
\item \strong{property}: a property of a node or relationship, access the full
list of properties: \verb{con <- getOption("con"); con$get_property_keys()}
\item \strong{relationship}: a relationship between nodes, access the full
list of relationships: \verb{con <- getOption("con"); con$get_relationships()}
\item Species information can see
\href{https://reactome.org/content/schema/objects/Species}{here}, or run
\code{View(matchObject(schemaClass = "Species"))[['databaseObject']]} to view a
full table
}
}
\examples{
# fetch instance by class
all.species <- matchObject(schemaClass = "Species")

# fetch instance by name
matchObject(displayName = "RCOR1 [nucleoplasm]", 
           returnedAttributes=c("stId", "speciesName"))

# fetch instance by id
## Reactome id
matchObject(id = "R-HSA-9626034")
## non-Reactome id
matchObject(id = "P60484", databaseName = "UniProt")

# fecth instances by relationship
matchObject(relationship="inferredTo", limit=10)

# fetch instances by property
property.list <- list(hasEHLD = TRUE, isInDisease = TRUE)
matchObject(property = property.list, 
    returnedAttributes = c("displayName", "stId", "isInDisease", "hasEHLD"), 
    limit=20)

}
\seealso{
\link{multiObjects} for multiple ids

Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchInteractors}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
