% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchPaperObjects}
\alias{matchPaperObjects}
\title{MATCH objects related to a paper}
\usage{
matchPaperObjects(
  pubmed.id = NULL,
  displayName = NULL,
  type = c("row", "graph")
)
}
\arguments{
\item{pubmed.id}{PubMed identifier of a paper}

\item{displayName}{paper title}

\item{type}{return results as a list of dataframes (\strong{'row'}),
or as a graph object (\strong{'graph'})}
}
\value{
Reactome instances associated with a paper
}
\description{
Fetch Reactome instances related to a paper by its PubMed id or title
}
\examples{
\dontrun{
# fetch Reactome instances by paper title
matchPaperObjects(displayName="Chaperone-mediated autophagy at a glance")

# fetch Reactome instances by pubmed id
matchPaperObjects(pubmed.id="20797626", type="graph")
matchPaperObjects(pubmed.id="23515720", type="row")
}
}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchInteractors}()},
\code{\link{matchObject}()},
\code{\link{matchPEroles}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
