% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchPrecedingAndFollowingEvents}
\alias{matchPrecedingAndFollowingEvents}
\title{MATCH the preceding/following Events}
\usage{
matchPrecedingAndFollowingEvents(
  event.id = NULL,
  event.displayName = NULL,
  species = NULL,
  depth = 1,
  all.depth = FALSE,
  type = c("row", "graph")
)
}
\arguments{
\item{event.id}{a stable or db id of an Event}

\item{event.displayName}{displayName of an Event}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{depth}{number of depths}

\item{all.depth}{if set to \code{TRUE}, all RLEs connected to the given Event in all depths returned}

\item{type}{to return results as a list of dataframes (\strong{'row'}) or as a graph object (\strong{'graph'})}
}
\value{
preceding/following Events connected to the given Event in specified depth(s), default depth = 1
}
\description{
To get Events connected with "precedingEvent" relationship
}
\examples{
matchPrecedingAndFollowingEvents("R-HSA-983150", all.depth=TRUE, type="row")
}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchObject}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
