% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchReactions}
\alias{matchReactions}
\title{MATCH Reactions}
\usage{
matchReactions(
  id = NULL,
  name = NULL,
  species = NULL,
  all.depths = FALSE,
  depths = NULL,
  type = c("row", "graph")
)
}
\arguments{
\item{id}{a stable or db id of a ReactionLikeEvent}

\item{name}{a ReactionLikeEvent name}

\item{species}{name or taxon id or dbId or abbreviation of species}

\item{all.depths}{if set to \code{TRUE}, all RLEs connected to the given RLE in all depths returned}

\item{depths}{number of depths}

\item{type}{to return the results as a list of dataframes or a graph object}
}
\value{
preceding/following RLEs connected to the given RLE in specified depth(s), default depth = 1
}
\description{
MATCH Reactions
}
\examples{
matchReactions("R-HSA-983150", all.depths=TRUE, type="row")
}
\concept{match}
