% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchReactionsInPathway}
\alias{matchReactionsInPathway}
\title{MATCH Reactions in associated Pathway}
\usage{
matchReactionsInPathway(
  event.id = NULL,
  event.name = NULL,
  species = NULL,
  type = c("row", "graph")
)
}
\arguments{
\item{event.id}{stId or dbId of Event}

\item{event.name}{name of Event}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{type}{return results as a list of dataframes (\strong{'row'}) or as a graph object (\strong{'graph'})}
}
\value{
Reactions related to the given Pathway/Reaction
}
\description{
MATCH Reactions in associated Pathway
}
\examples{
matchReactionsInPathway("R-HSA-1369062", type="graph")
matchReactionsInPathway("R-HSA-5682285", type="row")
}
\seealso{
Other match: 
\code{\link{matchHierarchy}()},
\code{\link{matchObject}()},
\code{\link{matchPrecedingAndFollowingEvents}()}
}
\concept{match}
