% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchReferrals}
\alias{matchReferrals}
\title{MATCH biological referrals}
\usage{
matchReferrals(
  id = NULL,
  displayName = NULL,
  main = TRUE,
  depth = 1,
  all.depth = FALSE,
  species = NULL,
  type = c("row", "graph")
)
}
\arguments{
\item{id}{stId or dbId of a Reactome object}

\item{displayName}{displayName of a Reactome object}

\item{main}{if set to \code{TRUE}, only \strong{first-class} referrals returned}

\item{depth}{number of depths}

\item{all.depth}{if set to \code{TRUE}, connected objects in all depths returned}

\item{species}{name or taxon id or dbId or abbreviation of a species}

\item{type}{return results as a list of dataframes (\strong{'row'}),
or as a graph object (\strong{'graph'})}
}
\value{
referrals of the given instance
}
\description{
This method retrieves Reactome objects that are connected with the given
object in a \emph{reverse} relationship. For example, to find Pathways containing
the given Reaction.
}
\details{
For now it just focuses on biological referrals in the following Classes:
"Event", "PhysicalEntity", "Regulation", "CatalystActivity","ReferenceEntity"
, "Interaction", "AbstractModifiedResidue".
}
\examples{
\dontrun{
matchReferrals("113454", type="row")
matchReferrals("R-HSA-112479", main=FALSE, all.depth=TRUE, type="row")
}
}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchInteractors}()},
\code{\link{matchObject}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReactionsInPathway}()}
}
\concept{match}
