% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{multiObjects}
\alias{multiObjects}
\title{Retrieve multiple Reactome objects}
\usage{
multiObjects(ids, databaseName = NULL, speedUp = FALSE, cluster = 2)
}
\arguments{
\item{ids}{Reactome stIds or dbIds, or non-Reactome ids}

\item{databaseName}{database name}

\item{speedUp}{if set to \code{TRUE}, doParallel method will be used}

\item{cluster}{the number of cluster in doParallel}
}
\value{
Reactome database objects for the given ids
}
\description{
Retrieve multiple Reactome objects
}
\examples{
# "ids" can be Reactome or non-Reactome ids
ids <- c("P02741", "P08887", "P08505", "Q9GZQ8", "Q9NZH6")
res <- multiObjects(ids, databaseName = "UniProt", speedUp = TRUE)

}
