% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnestListCol}
\alias{unnestListCol}
\title{Unnest a column of lists in a dataframe}
\usage{
unnestListCol(df, column = "properties")
}
\arguments{
\item{df}{dataframe where a column to be unnested}

\item{column}{specific column to be unnested}
}
\value{
an unnested dataframe for network visualization
}
\description{
Unnest a column of lists in a dataframe
}
\examples{
# nodes <- unnestListCol(graph$nodes, "properties")
}
