% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchHierarchy}
\alias{matchHierarchy}
\alias{matchInteractors}
\title{MATCH hierarchy}
\usage{
matchHierarchy(
  id,
  resource = "Reactome",
  species = NULL,
  type = c("row", "graph")
)

matchInteractors(pe.id, species = NULL, type = c("row", "graph"))
}
\arguments{
\item{id}{stId or dbId of Event/PhysicalEntity; or id of an external database}

\item{resource}{database name. All listed databases see \href{here}{https://reactome.org/content/schema/objects/ReferenceDatabase}}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{type}{return results as a list of dataframes (\strong{'row'}) or as a graph object (\strong{'graph'})}

\item{pe.id}{stId or dbId of PhysicalEntity}
}
\value{
hierarchical instances of the given id and resource

interactions of a given PhysicalEntity
}
\description{
Retrieve hierarchical data: Pathway-Reaction-Entity

Retrieve interactions of a given PhysicalEntity
}
\examples{
matchHierarchy(id="P04637", resource="UniProt", type="row")
matchHierarchy(id="R-HSA-196015", type="row")
matchHierarchy(id="R-HSA-1369062", type="graph")

matchInteractors(996766)
}
\seealso{
Other match: 
\code{\link{matchEventPaths}()},
\code{\link{matchObject}()},
\code{\link{matchReactionsInPathway}()}

Other match: 
\code{\link{matchEventPaths}()},
\code{\link{matchObject}()},
\code{\link{matchReactionsInPathway}()}
}
\concept{match}
