% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchHierarchy}
\alias{matchHierarchy}
\title{MATCH hierarchy}
\usage{
matchHierarchy(
  id = NULL,
  displayName = NULL,
  databaseName = "Reactome",
  species = NULL,
  type = c("row", "graph")
)
}
\arguments{
\item{id}{stId or dbId of an Event/Entity; or an external id}

\item{displayName}{displayName of Event/PhysicalEntity/ReferenceEntity}

\item{databaseName}{database name}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{type}{return results as a list of dataframes (\strong{'row'}),
or as a graph object (\strong{'graph'})}
}
\value{
hierarchical instances of the given id and databaseName
}
\description{
Reactome data are organized in a hierarchical way: Pathway-Reaction-Entity.
This function retrieves the hierarchical data of a given Event
(Pathway or Reaction) or Entity (PhysicalEntity or ReferenceEntity).
}
\examples{
## use the Reactome displayName of a UniProt object
uniprot.name <- "UniProt:P04637 TP53"
# matchHierarchy(displayName=uniprot.name, 
#                databaseName="UniProt", type="row")
# matchHierarchy(id="R-HSA-1369062", type="graph")
}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchInteractors}()},
\code{\link{matchObject}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
