% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchObject}
\alias{matchObject}
\title{Basic query for database object}
\usage{
matchObject(
  id = NULL,
  displayName = NULL,
  schemaClass = NULL,
  species = NULL,
  attribute = NULL,
  databaseName = "Reactome"
)
}
\arguments{
\item{id}{Reactome stId or dbId, or non-Reactome identifier}

\item{displayName}{displayName of a database object}

\item{schemaClass}{schema class of a database object}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{attribute}{specific attribute(s) to be returned. If set to \code{NULL}, all attributes returned}

\item{databaseName}{database name}
}
\value{
Reactome database object matched with the given id or name
}
\description{
Fetch instance by Reactome dbId/stId/displayName/schemaClass
or non-Reactome identifier/displayName
}
\examples{
# fetch instance by class
all.species <- matchObject(schemaClass = "Species")

# fetch instance by name
matchObject(displayName = "RCOR1 [nucleoplasm]", attribute=c("stId", "speciesName"))

# fetch instance by id
matchObject(id = "R-HSA-9626034")
matchObject(id = "P60484", species="pig", databaseName = "UniProt")
}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
