% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchObject}
\alias{matchObject}
\title{Basic query for database object}
\usage{
matchObject(
  id = NULL,
  displayName = NULL,
  schemaClass = NULL,
  species = NULL,
  returnAttribute = NULL,
  property = NULL,
  relationship = NULL,
  limit = NULL,
  databaseName = "Reactome"
)
}
\arguments{
\item{id}{Reactome stId or dbId, or non-Reactome identifier}

\item{displayName}{displayName of a database object}

\item{schemaClass}{schema class of a database object}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{returnAttribute}{specific attribute(s) to be returned. If set to \code{NULL}, all attributes returned}

\item{property}{a list of property keys and values}

\item{relationship}{a relationships type}

\item{limit}{the limit of returned objects}

\item{databaseName}{database name}
}
\value{
Reactome database object matched with the given conditions
}
\description{
Fetch instance by Reactome dbId/stId/displayName/schemaClass
or non-Reactome identifier/displayName
}
\examples{
# fetch instance by class
all.species <- matchObject(schemaClass = "Species")

# fetch instance by name
matchObject(displayName = "RCOR1 [nucleoplasm]", returnAttribute=c("stId", "speciesName"))

# fetch instance by id
## Reactome id
matchObject(id = "R-HSA-9626034")
## non-Reactome id
matchObject(id = "P60484", databaseName = "UniProt")

# fecth instances by relationship
matchObject(relationship="inferredTo", limit=10)

# fetch instances by property
property.list <- list(hasEHLD = TRUE, isInDisease = TRUE)
matchObject(property = property.list, 
            returnAttribute = c("displayName", "stId", "isInDisease", "hasEHLD"), 
            limit=20)

}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
