% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchObject}
\alias{matchObject}
\title{Basic query for database object}
\usage{
matchObject(id = NULL, name = NULL, species = "human")
}
\arguments{
\item{id}{stId or dbId of database object}

\item{name}{name of database object}

\item{species}{name or taxon id or dbId or abbreviation of specified species}
}
\value{
data of the given id or name
}
\description{
Fetch instance by dbId/stId/displayName
}
\examples{
matchObject(name = "RCOR1 [nucleoplasm]")
}
\seealso{
Other match: 
\code{\link{matchEventPaths}()},
\code{\link{matchHierarchy}()},
\code{\link{matchReactionsInPathway}()}
}
\concept{match}
