% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchPEroles}
\alias{matchPEroles}
\title{MATCH roles of PhysicalEntity}
\usage{
matchPEroles(
  pe.id = NULL,
  pe.displayName = NULL,
  species = NULL,
  type = c("row", "graph")
)
}
\arguments{
\item{pe.id}{stId or dbId of a PhysicalEntity}

\item{pe.displayName}{displayName of a PhysicalEntity}

\item{species}{name or taxon id or dbId or abbreviation of a species}

\item{type}{return results as a list of dataframes (\strong{'row'}),
or as a graph object (\strong{'graph'})}
}
\value{
information of the given PhysicalEntity and its role(s)
}
\description{
This function retrieves the role(s) of a given PhysicalEntity including:
\itemize{
\item Input
\item Output
\item Regulator
\item Catalyst
}
}
\examples{
\dontrun{
matchPEroles(pe.id = "R-HSA-8944354", type = "graph")
matchPEroles(pe.displayName = "2SUMO1:MITF [nucleoplasm]", 
             species = "pig", type = "row")
}
}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchInteractors}()},
\code{\link{matchObject}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
