% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchPrecedingAndFollowingEvents}
\alias{matchPrecedingAndFollowingEvents}
\title{MATCH the preceding/following Events}
\usage{
matchPrecedingAndFollowingEvents(
  event.id = NULL,
  event.displayName = NULL,
  species = NULL,
  depth = 1,
  all.depth = FALSE,
  type = c("row", "graph")
)
}
\arguments{
\item{event.id}{stId/dbId of an Event}

\item{event.displayName}{displayName of an Event}

\item{species}{name or taxon id or dbId or abbreviation of specified species}

\item{depth}{number of depths}

\item{all.depth}{if set to \code{TRUE}, all RLE(s) connected to the given Event
in all depths returned}

\item{type}{to return results as a list of dataframes (\strong{'row'}),
or as a graph object (\strong{'graph'})}
}
\value{
preceding/following Events connected to the given Event in specified
depth(s), default depth = 1
}
\description{
This method can find preceding and following ReactionLikeEvents (RLEs) of a
specific Event with the relationship 'precedingEvent'.
The argument "depth" is used to describe the "variable length relationships"
in Neo4j, default is 1 (i.e. immediately connected); or you can set
\code{all.depth = TRUE} to retrieve the whole context.
}
\examples{
\dontrun{
matchPrecedingAndFollowingEvents("R-HSA-983150", depth=2, type="row")
}
}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchInteractors}()},
\code{\link{matchObject}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchReactionsInPathway}()},
\code{\link{matchReferrals}()}
}
\concept{match}
