% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{matchReactionsInPathway}
\alias{matchReactionsInPathway}
\title{MATCH Reactions in associated Pathway}
\usage{
matchReactionsInPathway(
  event.id = NULL,
  event.displayName = NULL,
  species = NULL,
  type = c("row", "graph")
)
}
\arguments{
\item{event.id}{stId or dbId of an Event}

\item{event.displayName}{displayName of an Event}

\item{species}{name or taxon id or dbId or abbreviation of a species}

\item{type}{return results as a list of dataframes (\strong{'row'}),
or as a graph object (\strong{'graph'})}
}
\value{
Reactions connected to the given Pathway/Reaction via
'hasEvent' relationships
}
\description{
This method could find all Reactions connected with a given Pathway by
the relationship 'hasEvent'. Also, the input can be a Reaction, the result
would then be Pathway(s) linked via 'hasEvent' together with other Reactions
linked with the Pathways(s).
}
\examples{
matchReactionsInPathway("R-HSA-1369062", type="graph")
matchReactionsInPathway("R-HSA-5682285", type="row")
}
\seealso{
Other match: 
\code{\link{matchDiseases}()},
\code{\link{matchHierarchy}()},
\code{\link{matchInteractors}()},
\code{\link{matchObject}()},
\code{\link{matchPEroles}()},
\code{\link{matchPaperObjects}()},
\code{\link{matchPrecedingAndFollowingEvents}()},
\code{\link{matchReferrals}()}
}
\concept{match}
