% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{multiObjects}
\alias{multiObjects}
\title{Retrieve multiple Reactome objects}
\usage{
multiObjects(ids, databaseName = "Reactome", speedUp = FALSE, cluster = 2)
}
\arguments{
\item{ids}{Reactome stIds/dbIds, or non-Reactome ids}

\item{databaseName}{database name}

\item{speedUp}{set \code{TRUE} to use \code{\link[doParallel]{doParallel}} method}

\item{cluster}{the number of cluster in \code{\link[parallel]{makeCluster}}}
}
\value{
Reactome database objects for the given ids
}
\description{
The \link{matchObject} function takes only one id/name at a time, this method
allows you to input many ids and get an aggregated table for their detailed
information. It can only accept \strong{ids} for now.
}
\examples{
\dontrun{
# "ids" can be Reactome or non-Reactome ids
ids <- c("P02741", "P08887", "P08505", "Q9GZQ8")
res <- multiObjects(ids, databaseName="UniProt", speedUp=TRUE)
}
}
\seealso{
\link{matchObject} for details
}
