% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphdb4r.R
\docType{package}
\name{reactome.graphdb4r-package}
\alias{reactome.graphdb4r}
\alias{reactome.graphdb4r-package}
\title{reactome.graphdb4r: R Interface for Reactome Graph Database}
\description{
This package is to provide an interface to query data with network structure
 in the Reactome Graph Database. There is also another package, reactome4r, for queries in 
 Reactome Content Service.
}
\author{
\strong{Maintainer}: Chi-Lam Poon \email{clpoon807@gmail.com}

}
\keyword{internal}
