% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphdb4r-package.R
\docType{package}
\name{ReactomeGraph4R-package}
\alias{ReactomeGraph4R}
\alias{ReactomeGraph4R-package}
\title{ReactomeGraph4R: Interface for the Reactome Graph Database}
\description{
Pathways, reactions, and biological entities in Reactome knowledge 
    are systematically represented as an ordered network and stored in the 
    Reactome Graph Database. This package provides an interface to query 
    interconnected data in the local Neo4j database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reactome/ReactomeGraph4R}
  \item Report bugs at \url{https://github.com/reactome/ReactomeGraph4R/issues}
}

}
\author{
\strong{Maintainer}: Chi-Lam Poon \email{clpoon807@gmail.com} (\href{https://orcid.org/0000-0001-6298-7099}{ORCID})

Authors:
\itemize{
  \item Robin Haw \email{Robin.Haw@oicr.on.ca}
}

Other contributors:
\itemize{
  \item Reactome \email{help@reactome.org} [copyright holder]
}

}
\keyword{internal}
