% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphdb4r-package.R
\docType{package}
\name{ReactomeGraph4R-package}
\alias{ReactomeGraph4R}
\alias{ReactomeGraph4R-package}
\title{ReactomeGraph4R: R Interface for Reactome Graph Database}
\description{
This package is to provide an interface to query data with the network structure
    in Reactome Graph Database. There is also another package, ReactomeContentService4R, for queries in 
    Reactome Content Service.
}
\author{
\strong{Maintainer}: Reactome \email{help@reactome.org}

Authors:
\itemize{
  \item Chi-Lam Poon \email{clpoon807@gmail.com} (\href{https://orcid.org/0000-0001-6298-7099}{ORCID})
}

}
\keyword{internal}
